# backend/tests/snapshots/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swings API",
    "description": "Swings public + admin HTTP API. Mutating endpoints are annotated; GETs are still fully callable but not listed here (see FDN-02 scope).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative to deploying host"
    }
  ],
  "paths": {
...
    "/api/admin/pricing/plans": {
      "post": {
        "tags": [
          "pricing"
        ],
        "operationId": "admin_create_plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePricingPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Plan created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingPlan"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/pricing/plans/{id}": {
      "put": {
        "tags": [
          "pricing"
        ],
        "operationId": "admin_update_plan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Plan id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePricingPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Plan updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingPlan"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Plan not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "pricing"
        ],
        "operationId": "admin_delete_plan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Plan id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan deleted"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Plan not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/pricing/plans/{id}/toggle": {
      "post": {
        "tags": [
          "pricing"
        ],
        "operationId": "admin_toggle_plan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Plan id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan active flag toggled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingPlan"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Plan not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePricingPlanRequest": {
        "type": "object",
        "required": [
          "name",
          "amount_cents"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "stripe_price_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "stripe_product_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "amount_cents": {
            "type": "integer",
            "format": "int32"
          },
          "currency": {
            "type": [
              "string",
              "null"
            ]
          },
          "interval": {
            "type": [
              "string",
              "null"
            ]
          },
          "interval_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "trial_days": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "features": {},
          "highlight_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_popular": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "is_active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "sort_order": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
      "PricingPlan": {
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "amount_cents",
          "currency",
          "interval",
          "interval_count",
          "trial_days",
          "features",
          "is_popular",
          "is_active",
          "sort_order",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "stripe_price_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "stripe_product_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "amount_cents": {
            "type": "integer",
            "format": "int32"
          },
          "currency": {
            "type": "string"
          },
          "interval": {
            "type": "string"
          },
          "interval_count": {
            "type": "integer",
            "format": "int32"
          },
          "trial_days": {
            "type": "integer",
            "format": "int32"
          },
          "features": {},
          "highlight_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_popular": {
            "type": "boolean"
          },
          "is_active": {
            "type": "boolean"
          },
          "sort_order": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdatePricingPlanRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "slug": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "stripe_price_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "stripe_product_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "amount_cents": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "currency": {
            "type": [
              "string",
              "null"
            ]
          },
          "interval": {
            "type": [
              "string",
              "null"
            ]
          },
          "interval_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "trial_days": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "features": {},
          "highlight_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_popular": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "is_active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "sort_order": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}